// apps/frontend/src/lib/tour/types.ts
export interface PopoverDOM {
  wrapper: HTMLElement
  arrow: HTMLElement
  title: HTMLElement
  description: HTMLElement
  footer: HTMLElement
  progress: HTMLElement
  previousButton: HTMLButtonElement
  nextButton: HTMLButtonElement
  closeButton: HTMLButtonElement
  footerButtons: HTMLElement
}